add review endpoints

+++ SportZoneServer.Common/Responses/Product/ProductResponse.cs
@@ -16,7 +16,7 @@ public class ProductResponse
     
     public decimal DiscountedPrice { get; set; }
     
-    public byte Rating { get; set; } 
+    public double Rating { get; set; } 
 
     public uint Quantity { get; set; }
